refactoring

+++ src/main/java/forex/entity/Order.java
@@ -11,7 +11,7 @@ public class Order {
     private Strategy strategy;
     private ClosePriceType closePriceType;
     private LocalDateTime openTime;
-    private double openPrice;
-    private double closePrice;
-    private double profit;
+    private int openPrice;
+    private int closePrice;
+    private int profit;
 }
